--- src/Animals/Dieable.java
@@ -1,9 +1,0 @@
-package Animals;
-
-public interface Dieable {
-    private void die () {
-    }
-    private void starvation () {
-    }
-
-}
