--- src/Animals/Eatable.java
@@ -1,5 +1,0 @@
-package Animals;
-
-public interface Eatable {
-    void eat(int amountOfFood);
-}
