--- src/Animals/Herbivores/Deer.java
@@ -1,16 +1,0 @@
-package Animals.Herbivores;
-
-import Animals.AnimalCharacteristics;
-import Animals.AnimalType;
-import Island.Island;
-
-public class Deer extends Herbivores {
-    public Deer(Island island, AnimalCharacteristics animalCharacteristics) {
-        super(island, animalCharacteristics);
-    }
-
-    @Override
-    public AnimalType getAnimalType () {
-        return AnimalType.DEER;
-    }
-}
