--- src/Animals/Herbivores/Mouse.java
@@ -1,17 +1,0 @@
-package Animals.Predators;
-
-import Animals.AnimalCharacteristics;
-import Animals.AnimalType;
-import Animals.Herbivores.Herbivores;
-import Island.Island;
-
-public class Mouse extends Herbivores {
-    public Mouse(Island island, AnimalCharacteristics animalCharacteristics) {
-        super(island, animalCharacteristics);
-    }
-
-    @Override
-    public AnimalType getAnimalType() {
-        return AnimalType.MOUSE;
-    }
-}
