--- src/Animals/Herbivores/Rabbit.java
@@ -1,16 +1,0 @@
-package Animals.Herbivores;
-
-import Animals.AnimalCharacteristics;
-import Animals.AnimalType;
-import Island.Island;
-
-public class Rabbit extends Herbivores {
-    public Rabbit(Island island, AnimalCharacteristics animalCharacteristics) {
-        super(island, animalCharacteristics);
-    }
-
-    @Override
-    public AnimalType getAnimalType() {
-        return AnimalType.RABBIT;
-    }
-}
