--- src/Animals/Predators/Predators.java
@@ -1,12 +1,0 @@
-package Animals.Predators;
-
-import Animals.Animal;
-import Animals.AnimalCharacteristics;
-import Island.Island;
-
-public abstract class Predators extends Animal {
-    public Predators(Island island, AnimalCharacteristics animalCharacteristics) {
-        super(island, animalCharacteristics);
-    }
-}
-
