--- src/Animals/Predators/Wolf.java
@@ -1,16 +1,0 @@
-package Animals.Predators;
-
-import Animals.AnimalCharacteristics;
-import Animals.AnimalType;
-import Island.Island;
-
-public class Wolf extends Predators {
-    public Wolf(Island island, AnimalCharacteristics animalCharacteristics) {
-        super(island, animalCharacteristics);
-    }
-
-    @Override
-    public AnimalType getAnimalType() {
-        return AnimalType.WOLF;
-    }
-}
