--- src/Config/Config.java
@@ -1,4 +1,0 @@
-package Config;
-
-public class Config {
-}
