--- src/Island/Direction.java
@@ -1,5 +1,0 @@
-package Island;
-
-public enum Direction {
-    UP, DOWN, LEFT, RIGHT;
-}
