--- src/Island/Plants.java
@@ -1,13 +1,0 @@
-package Island;
-
-public class Plants {
-
-//    private static final int GROW_FACTOR = 2;
-
-    private int plantsLevel;
-
-    public Plants(int plantsLevel) {
-        this.plantsLevel = plantsLevel;
-    }
-
-}
