+++ src/animals/AnimalType.java
@@ -1,0 +1,32 @@
+package animals;
+
+import java.util.List;
+
+public enum AnimalType {
+// Herbivores
+    HORSE,
+    DEER,
+    RABBIT,
+    MOUSE,
+    GOAT,
+    SHEEP,
+    WILD_BOAR,
+    BUFFALO,
+    DUCK,
+//Predators
+    FOX,
+    WOLF,
+    SNAKE,
+    BEAR,
+    EAGLE;
+
+
+    public static List<AnimalType> getHerbivoresTypes() {
+        return List.of(HORSE,DEER,RABBIT,MOUSE,GOAT,SHEEP,WILD_BOAR,BUFFALO,DUCK);
+    }
+
+    public static List<AnimalType> getPredatorsTypes() {
+        return List.of(WOLF,SNAKE,BEAR,EAGLE);
+    }
+
+}
