+++ src/animals/Dieable.java
@@ -1,0 +1,9 @@
+package animals;
+
+public interface Dieable {
+    private void die () {
+    }
+    private void starvation () {
+    }
+
+}
