+++ src/animals/Eatable.java
@@ -1,0 +1,5 @@
+package animals;
+
+public interface Eatable {
+    void eat();
+}
