+++ src/animals/predators/Fox.java
@@ -1,0 +1,16 @@
+package animals.predators;
+
+import animals.AnimalCharacteristics;
+import animals.AnimalType;
+import island.Island;
+
+public class Fox extends Predators {
+    public Fox(Island island, AnimalCharacteristics animalCharacteristics) {
+        super(island, animalCharacteristics);
+    }
+
+    @Override
+    public AnimalType getAnimalType() {
+        return AnimalType.FOX;
+    }
+}
