+++ src/animals/predators/Snake.java
@@ -1,0 +1,16 @@
+package animals.predators;
+
+import animals.AnimalCharacteristics;
+import animals.AnimalType;
+import island.Island;
+
+public class Snake extends Predators {
+    public Snake(Island island, AnimalCharacteristics animalCharacteristics) {
+        super(island, animalCharacteristics);
+    }
+
+    @Override
+    public AnimalType getAnimalType() {
+        return AnimalType.SNAKE;
+    }
+}
