+++ src/config/Config.java
@@ -1,0 +1,4 @@
+package config;
+
+public class Config {
+}
