+++ src/island/Direction.java
@@ -1,0 +1,5 @@
+package island;
+
+public enum Direction {
+    UP, DOWN, LEFT, RIGHT;
+}
