+++ src/island/Plants.java
@@ -1,0 +1,13 @@
+package island;
+
+public class Plants {
+
+//    private static final int GROW_FACTOR = 2;
+
+    private int plantsLevel;
+
+    public Plants(int plantsLevel) {
+        this.plantsLevel = plantsLevel;
+    }
+
+}
